Change tokens from i64 to i32 (#68)
BUG=b/348016917

+++ ai_edge_torch/generative/examples/stable_diffusion/clip.py
@@ -60,8 +60,8 @@ class CLIP(nn.Module):
     )
 
   @torch.inference_mode
-  def forward(self, tokens: torch.LongTensor) -> torch.FloatTensor:
-    tokens = tokens.type(torch.long)
+  def forward(self, tokens: torch.IntTensor) -> torch.FloatTensor:
+    tokens = tokens.type(torch.int)
 
     state = self.tok_embedding(tokens) + self.tok_embedding_position
     for layer in self.transformer_blocks:
